requirements_cloud.txt
streamlit==1.35.0
easyocr
opencv-python-headless
Pillow==10.3.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
scipy==1.13.0